+++ src/Services/Report/Report.API/Services/DocumentService.cs
@@ -4,16 +4,41 @@ using OfficeOpenXml;
 using Report.API.Infrastructure.Data;
 using MongoDB.Driver.Linq;
 using Report.API.Models;
+using MassTransit;
+using EventBus.IntegrationEvents;
 
 namespace Report.API.Services;
 
 public class DocumentService : IDocumentService
 {
     private readonly ReportContext _context;
+    private readonly IBus _bus;
 
-    public DocumentService(ReportContext context)
+    public DocumentService(ReportContext context, IBus bus)
     {
         _context = context;
+        _bus = bus;
+    }
+
+    public async Task<string> MakeReportRequestAsync()
+    {
+        var requestedAt = DateTime.UtcNow;
+
+        var report = new Entities.Report
+        {
+            RequestedAt = requestedAt,
+            Status = ValueObjects.ReportStatus.InProgress
+        };
+
+        await _context.Reports.InsertOneAsync(report);
+
+        await _bus.Publish(new ReportRequestedEvent()
+        {
+            ReportId = report.Id,
+            RequestedAt = requestedAt,
+        });
+
+        return report.Id;
     }
 
     public async Task<string> CreateExcelFileAsync<T>(IEnumerable<T> data, string fileDirectory)
@@ -33,6 +58,7 @@ public class DocumentService : IDocumentService
     public async Task<IEnumerable<Entities.Report>> GetLocationReportsAsync()
     {
         var reports = await _context.Reports.AsQueryable().ToListAsync();
+
         return reports;
     }
