+++ src/Services/Report/Report.API/Services/IDocumentService.cs
@@ -6,4 +6,5 @@ public interface IDocumentService
     Task<string> CreateExcelFileAsync<T>(IEnumerable<T> data, string fileDirectory);
     Task<FileStream> GetGeportFileAsync(string id);
     Task<IEnumerable<Entities.Report>> GetLocationReportsAsync();
+    Task<string> MakeReportRequestAsync();
 }
