--- src/Services/Report/Report.API/Services/ILocationService.cs
@@ -1,7 +1,0 @@
-﻿
-namespace Report.API.Services;
-
-public interface ILocationService
-{
-    Task<string> MakeReportRequestAsync();
-}
